backend/requirements.txt
flask>=3.1.0
flask-cors>=5.0.0
google-cloud-bigquery>=3.20.0
gunicorn>=23.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
pyarrow>=15.0.0
